Reverted build.js change and renamed test script
